HAAAAAANDS
I went on a coding bender and never committed. But MushROMs is still in
early development so tracking crucial changes in this stage isn't the
most important thing to me. Overall, I made a few attempts to improve
some code, fix some bugs, and add new features. Everything is still in
development.